# Performance: Lazy item index