--- a/libs/pavex/Cargo.toml
+++ b/libs/pavex/Cargo.toml
@@ -38,3 +38,4 @@ toml_edit = { version = "0.19.8", features = ["serde"] }
 xdg-home = "1.0.0"
 semver = "1.0.17"
 rusqlite = { version = "0.29.0", features = ["bundled"] }
+bincode = "1"