… libs/Cargo.lock
… libs/pavex/src/rustdoc/compute/cache.rs
… libs/pavex/src/rustdoc/queries.rs