The `index` field in `rustdoc`'s JSON output can be massive and take a **significant** amount of time to deserialize. E.g. 0.18s for `core`.
We are doing a lot of unnecessary work considering that we access a tiny portion of the index in most cases.

This PR builds on top of our SQLite cache to avoid this cost, when possible.
We load into memory a byte array and an `id <> delimiter` map that tells us where each item is inside the byte buffer. We then deserialize it on the fly when `pavex` needs to access it.

After this change, loading `core` from the cache takes ~0.02s, almost 10x faster than before.

Code-generation for our `skeleton` example, with a warm cache, goes from 1.19s to 0.83s. Not bad!